+++ lib/Fhaculty/Graph/Algorithm/Degree.php
@@ -87,4 +87,22 @@ class Degree extends Base
 
         return false;
     }
+
+    /**
+     * checks whether the indegree of every vertex equals its outdegree
+     *
+     * @return boolean
+     * @uses Vertex::getDegreeIn()
+     * @uses Vertex::getDegreeOut()
+     */
+    public function isBalanced()
+    {
+        foreach ($this->graph->getVertices() as $vertex) {
+            if ($vertex->getDegreeIn() !== $vertex->getDegreeOut()) {
+                return false;
+            }
+        }
+
+        return true;
+    }
 }