+++ CHANGELOG.md
@@ -8,7 +8,7 @@ you spot any mistakes.
 
 * BC break: `Graph::createVertices()` now returns an array of vertices instead of the chainable `Graph` (#19)
 * BC break: Move `Set::getWeight()`, `Set::getWeightFlow()`, `Set::getWeightMin()` and `Set::isWeighted()` to new `Algorithm\Weight` (#33)
-* BC break: Move `Graph::getDegree()`, `Graph::getDegreeMin()`, `Graph::getDegreeMax()` and `Graph::isRegular()` to new `Algorithm\Degree` (#29)
+* BC break: Move `Graph::getDegree()`, `Graph::getDegreeMin()`, `Graph::getDegreeMax()`, `Graph::isRegular()` and `Graph::isBalanced()` to new `Algorithm\Degree` (#29)
 * Feature: `Graph::createVertices()` now also accepts an array of vertex IDs (#19)
 * Fix: Various issues with `Vertex`/`Edge` layout attributes (#32)